api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
numpy==2.0.2
scikit-learn==1.6.1
tensorflow==2.19.0
joblib==1.5.3
dill
python-dotenv
